Added first admin pireps implementation

+++ db/migrate/20240801135343_create_pireps.rb
@@ -7,7 +7,7 @@ class CreatePireps < ActiveRecord::Migration[7.1]
       t.string :departure_icao, null: false
       t.string :arrival_icao, null: false
       t.integer :flight_time_minutes, null: false
-      t.decimal :fuel_used, null: false
+      t.integer :fuel_used, null: false
       t.integer :cargo, null: false
       t.string :remarks, limit: 500
       t.string :status, default: 'pending'
